Fix preview delay hack
All logic for handling preview delays has been moved within the
`PreviewPlayer` struct. Instead of being paired together¸ the
fading in and out of previews are now handled separately. When the
song is changed while a preview is playing, both the fade-out timer
and the delay timer begin ticking. Once finished fading out, the fade-out timer
will remain like that, while the delay timer goes back to zero each time
a new chart is selected. When the delay timer finished, the fade-in timer
begins, fading in the track.

This commit configures a 500ms delay before a preview begins to play.