Support dedicated preview files
KSH charts can use the field `previewfile` to specify the name
of the file to be played as a preview. This ignores the offset and
duration parameters and plays the file from start to finish.

My rationale for this is that many rhythm games, both
community-driven and commercial, choose to store their preview
files this way instead of as offsets into the main audio. In the
interest of allowing charters to more easily work with charts
using files from those games, USC should support such a technique.